move endpoint_status_inspect to endpointedge package
